subprocess: set environ for breakout subprocess
When executing a process through the subprocess helper, we still need to
ensure some basic params are set. This promotes the setting of these
variables to the spawn vfunc so they apply to both gsubprocess/breakout.

Additionally, we rely on the chain ordering so that flatpak subprocess
launcher can set /app/bin into the path without this overriding.