phpize: add phpize plugin for writing PHP extensions
For the general purpose PHP extension case, where you're writing some C
or C++ code for an extension.

phpize is run to bootstrap the project, configure to configure, make to
build, make clean to clean and make install to install.

It looks for "config.m4" in the root of the project to determine that the
project is a phpize project.