ctags: perform more incremental ctags updates
This changes how we do ctags indexes to be a bit more friendly on larger
codebases. Now, we do a single index pass at startup of the project to
update all our ctags indexes. Then, when a file is saved, we only reindex
that directory, ignoring the other directories.

When reloading the indexes, we only pop the cache if the mtime is newer
so that old indexes shouldn't get reloaded/replaced.

In practice, if i save a file like src/foo.c, only src/tags will get
reloaded. Note that src/tags is actually based on the tags directory
in ~/.cache/gnome-builder/tags/$project_id, not the actual source tree.

I think there is a bunch more we can still improve, but this should make
things a bit better.