rustup: run rustup on the host system
We do all our rust toolchain management on the host, which means that we
don't have access to the rust toolchain inside the flatpak application
mount namespace. So ensure we run that on the host too using the bash
trick "bash -- /dev/stdin arguments" style and writing the script to
stdin.