flatpak: Monitor the right file after a rename event
When you save an edited file in Builder it renames a temporary file to
the appropriate file name, so we need to deal with that behavior
appropriately for the GFileMonitor that watches the flatpak manifest.
This commit fixes a bug where we were monitoring the temporary file
rather than the actual manifest after such a rename event.