flatpak: Keep file monitor for deleted manifests
Some git operations such as "git revert" cause the following sequence of
GFileMonitor events: DELETED, CREATED, CHANGED, CHANGES_DONE_HINT. This
means that we need to keep the monitor for a "deleted" manifest in order
to get the following events and recreate it. Unfortunately this also
means we're parsing the manifest and creating a config for it twice for
one set of changes, but not much can be done about that.