pipeline: add IdeBuildStage:check-stdout
This property is used to notify the pipeline that it needs to look at
stdout to extract errors using the error regexes. Typical build systems
such as make, will pass stderr on to the caller. However, others like
Ninja coalesce everything into a single stream, complicating the matter.

This means we burn extra overhead parsing stdout, but in practice, the
output we get from ninja is small enough that it should be roughly
equivalent.