Retab: convert indetation between tabs and spaces
When pasting code with diffrent indentation style then the one in the
Indentation settings, the indentation style of the inserted code stays the
same.

A new plugin retab has been added. The retab is accessible through the
Selection in the popup menu and converts the indentation style base on the
Indentation setting of the file.

https://bugzilla.gnome.org/show_bug.cgi?id=766266