gdb: port plugin to C
Python was fine for some prototyping, but the G-I issues are slowing down
my ability to iterate quickly. So I guess we'll just do this plugin in C.

I was hoping to use this one as an example, but I guess we'll just do a
mock debugger to show people how to do it.