flatpak: Add support for "simple" buildsystem
This commit adds support for flatpak's "simple" buildsystem, which just
runs the build commands in the manifest. The commands are loaded in the
IdeBuildPipeline initalization so they are available for any buildsystem
(and could conceivably be used with a non-flatpak config). Then the
"directory" build system is used since nothing more needs to be done
beyond running those commands.

https://bugzilla.gnome.org/show_bug.cgi?id=780674