greeter: Add build system tag to projects
In Builder projects can be opened using different build systems
depending on the project file that is chosen, which means the same
project can appear twice in the "Recent Projects" list of the greeter,
with no apparent difference. This commit adds a tag to each project so
the user knows what build system it uses, in addition to the tags
showing the language(s).