sourceview: add ctrl+shift+space to show references to symbol
This currently only works with Rust, as it is the only IdeSymbolResolver
that implements this vfunc. It should not be terribly difficult to
implement it for clang, but would be limited to files that have been
parsed and units loaded.

Realistically for other languages such as clang, we'll want a code index
to provide this stuff w/ background indexing.

We still need to make it easy to allow you to jump to the location, provide
project path relative filenames, and background load the symbol that is
found at that scope so that we can provide better labels.