xml-pack: add more internal nodes to the trees
Previously, we were integrating only needed/visible nodes
(the ones we need for the symbol tree view)

As a start, we integrate all start elements.
The final purpose is the be able to match a position
in the source doc with a node in the tree and
compare the doc tree with schemas tree for the yet to come
autocompletion based on schemas.