sourceview: move by visual column
Change movements to stick to the visual column instead of the line
offset. This make movements between lines with different (or buggy)
indentation more coherent with what the user expects.

Column index 0 is the first column which is shown as column 1 to the
user (this reflects the previous offset behavior).

* Convert saved line offset instances to columns
* Make ide_source_view_get_visual_column a wrapper for GtkSourceView's
  gtk_source_view_get_visual_column
* Use multiples of tab_width as tab stops. This might not universally hold
  in the future but it's a good start.
* ide_source_view_get_iter_at_visual_column is based on the column
  calculation of GtkSourceView's gtk_source_view_get_visual_column,
  including the FIXME comment about invisible characters.

IdeSourceLocation positions remain in character offsets because
tab_width of different files might be different.

https://bugzilla.gnome.org/show_bug.cgi?id=781342