plugins: allow ~/.local/share/gnome-builder/plugins under flatpak
The XDG_USER_DATA_DIR will be different when running under flatpak, but it
is considerably annoying to have to tell users where to put their plugins
based on how they got Builder.

Since Builder is a developer tool with access to $HOME anyway, we might as
well just try to do the right thing and give access to the documented
location even when running with Flatpak.