run-manager: Disable run button when can-build is FALSE
We're already disabling the build buttons when the build isn't possible
(e.g. due to a missing runtime) so it makes sense to disable the Run
button too (which does an install before running). Since the actions'
enabled state also depends on whether the IdeRunManager is busy or not,
we have to use a callback rather than a simple property binding. And
since the context isn't available in the init() we have to get it in the
async init instead.

https://bugzilla.gnome.org/show_bug.cgi?id=780688