transfer: fail transfer stages when on metered connections
If we are on a metered connection, then we don't want to be downloading
things like SDKs and such. Default to setting disable-when-metered to
TRUE. Note that if the ::query() phase of the build stage marks the
transfer as completed, then the stage will not fail.

A followup commit will contain the ability to configure if you want to
download on metered connections.

https://bugzilla.gnome.org/show_bug.cgi?id=781450