gsettings: make downloads over metered connections configurable
Some people may always be on metered connections, or specifically want to
download updates while transiently on a metered connection. This allows
the user to override the setting and do that.

In some ways, I wish we would provide UI flow when the problem exists, but
this is a stop-gap until we think about how we'd even go about doing that.
I suspect it's more effort than it's worth.

https://bugzilla.gnome.org/show_bug.cgi?id=781450