highlighter: add note about desired frame clock usage
If the highlighter had access to a GdkFrameClock, it would be nice to use
that to drive the buffer scanning. We could increase our quanta and be
more sure that we would not stomp on the next frame.

Also, lower the priority just a bit more so other callbacks fire before
our scanning work (we use G_PRIORITY_LOW all over the place).