buffer: set the initial state to loading
This allows us to set the buffer state to loading and then ack that state
once the BUFFER_LOADED signal is emitted in the buffer manager. This helps
ensure that we have a consistent flow for this. And doing that makes it
easier to disable various features while loading the buffer.