buffers: fix tracking of failed buffer loads
If we failed loading a buffer properly, such as by encoding error, we'll
always try to reload it at startup because it will never be removed from
the drafts. That can cause quite a bit of pain to recover since you'd have
to remove the drafts directory.

This ensures that we track the buffer, so that we can later clean it up.

It also moves some things around so that we always do the cleanup work
from buffer-unloaded's default handler.