bug-buddy: fork() and execv() gdb on SIGSEGV
This can be extremely helpful in situations where it is hard to
get a debugger in the normal process. We connect a SIGSEGV
handler that can safely fork()/execv() a new process (meaning
we can't use malloc, locks, etc).

That gdb process will inspect our process while we wait for it
to complete. After the gdb process exits, our crashing process
will allow itself to exit.