workbench: Introduce focus mode
When using a long-running application like GNOME Builder,
it is common that the user needs to put all efforts and
attention into it.

For these cases, having even less window chrome is desired,
and this can be achieved by making the window fullscreen
and automatically showing or hiding the header bar based
on the mouse position. This is the behavior of various
GNOME apps, like Gedit and Totem.

This commit, then, introduces this focus mode where the
window turns fullscreen and the header bar, autohidden.
This mode can be toggled by pressing F11, or by typing
'toggle-focus-mode' at the command bar.