todo: port todo plugin to C and use new sidebar
This ports the TODO plugin to C as I was tired of tracking down crashes in
PyGObject there (it's not necessarily to blame, but it was sufficiently
difficult to find the real culprite when python threading is in play).

Additionally, it tries to be more efficient with string storage than the
previous version by using a large GByte buffer with \0 terminated strings
inside that buffer. This allows us to have a single contiguous string
buffer (and that is good for the allocator).

Additionally, this switches to using the new IdeEditorSidebar.

I decided to not take the time to implement line-wrapping like the current
mockup has because doing that with CellRendererText is a bit obtuse. We
previously tried to use GtkListStore for stuff like this (where we would
get line wrapping for free) but it just wasn't tennable for lists over
about 100. So this sticks with the performance model of GtkTreeView and
just tries to get things close.