devhelp: move devhelp documentation to layout stack header
This helps us get away from the right sidebar being essential
to using Builder. Devhelp was the last panel there and can
naturally fit into the same sort of flow as the editor+symbols.

The use-case here is that the user brings up a devhelp view on
the right-hand of the split. Then that split can be used to
browse documentation (through the header) or automatically using
keyboard shortcuts from the editor.

The keyboard shortcuts for the editor still needs additional
plumbing to locate a view in a sibling stack.