find-other-file: add simple find-other-file plugin
This was previously implemented as an action to the editor view and only
knew how to cycle between a couple file types. This is more generic and
only looks at file prefixes.

We will still need some work on the interaction and HUD design, but this
contains the basic mechanics of what needs to be done.