layout: avoid global CSS style invalidation
Previously, we were invalidating the CSS on the global CSS
providers for the screen. That cascaded into lots of damages
and tanked our animation performance.

Instead, if we selectively add our CSS selector to widgets in
the layout stack, we can avoid lots of damages and things stay
quick and snappy.

It feels a bit evil, but hey, we wanna go fast.