indenter: add fallback when no indenter is supplied
We can have a very simple indenter as a fallback for the
interface. Just pass NULL for @self in those cases.

This tries to mimic GtkSourceView in that it just copies the
previous lines indentation (when possible). No indentation
is performed if we land on the first line of the buffer.

https://bugzilla.gnome.org/show_bug.cgi?id=784633