editor: hook up search bar
This starts hooking up the search bar. It is different than
previous in that we are using a new settings/context at the
editor layer rather than from the SourceView so that we can
easily destroy objects when not in use.

There is still a lot of work that can go on to refactor all of
this, but it is a start.