html-preview: add shortcut for html preview
This adds an ctrl+alt+p shortcut like we used to have a long
time ago. However, due to a bug in the shortcut controller
dispatch in libdazzle, it does not current activate unless it
is set to global (which we do not want).

We'll have fixes coming to libdazzle shortly.

Additionally, it would be nice to register the command with
the shortcut manager so that we can get automated UI in the
shortcuts window, but we need to figure out how to do that
from Python first.