source-view-mode: make value copies when emitting actions
I've noticed some weird reference issues when emitting signals
through the IdeSourceViewMode, so this tries to be safer by
making full references to the objects during the emission
procedure.

Additionally, we spill the value array copy to the stack by
using alloca since there should only ever be a small number of
values to copy.