indenter: only copy indent up to line end
When we do not have an indenter, we try to copy the indentation
style of gedit. We simply copy the previous line. However, we
were copying too much information by not stoping at the line
end.

This caused multiple newline insertions matching up to the
number of new lines after the insertion cursor (plus the newline
we added).

https://bugzilla.gnome.org/show_bug.cgi?id=785213