buildui: add sources directory to configuration view
This isn't tied to a configuration, but it is useful to see
here alongside build system (which isn't currently tied to a
configuration either).

I can see both the sources dir and build system being
configurable someday if we get around to supporting reloading of
the build system and alternate branching workflows.