symbol-tree: always expand the tree
By default, the tree nodes are builded lazyly and async using
a signal dispatched by the main loop, this make it hard
to use gtk_tree_view_expand_all.

We use the new construct-time "always-expand" DzlTree property
to provide easily the same effet as we always want the symbol tree
to be expanded.
.