buffermanager: protect against double open race
If the user double clicks in the project-tree, we could get into a race
condition where the first buffer is not yet registered. This adds a loading
hashtable so we can check if the buffer is being loaded.

If so, we ignore the second request by synthesizing an error upfront.