plugins: release plugins during IdeExtensionSetAdapater.dispose
We were leaking extensions when we were disposed without first NULL'ing
out the matching extensions. This ensures that we cleanup any future
extensions in the dispose phase.

To avoid breaking expectations about having a valid self->extensions
hashtable, we replace a dummy hashtable in case the object tries to
outlive the GObjectClass.dispose call.