debugger: add minimal debugger API
This is an initial implementation to provide a debugger API to Builder.  It is
not complete, but should be considered a technology preview as we complete the
feature-set that users come to expect from debugger tooling.

It should, however, give you an idea of where we are going with things, even if
UI and APIs should change a bit in the upcoming months.

Debuggers should implement the IdeDebugger API (which is a base class to make
things more convenient). A reference implementation using GDB will be provided
in a forth-coming commit.

Most operations are asynchronous, as debuggers can be very slow given some
hardware configurations.

Of note, we do not yet have support for diving into varibles such as fields,
variable hover, or data visualizers. These features and more avanced ones such
as server-side variables will come in the form of a future release, such as
GNOME 3.28.