line-changes: avoid GtkTextIter in line change interface
Looking up a GtkTextIter for a given line is actually quite expensive,
even with a BTree for lines. We don't actually need the GtkTextIter anyway
since all we care about is the line number (and the change monitors already
have access to the buffer to convert that into a GtkTextIter).

This shaves off a few % when rendering the gutters.