editor: implement utilities panel
This implements the utilities panel instead of just our "fallback" bottom
section that we had before.

It had a bit of styling for our two primary themes, Adwaita and Arc. The
normal variant of Adwaita has some tweaks to look flat, similar to the
mockups, but that has not yet been done for the dark mode (which should
maybe be done as well). It could probably use some :backdrop tweaks as
well.

This switches to using icons+tooltip text for the group sections.