debugger: add hover state to gutter
We still want a new omnigutter for 3.28, but this will get us by until
then. This does the simplist thing we can here and shows the breakpoint
icon when hovered. We would at least like some alpha, but I'm hesitant
to invest too much time in it when we'll throw it all way for the new
gutter.