codeindex: make code index service pausable
Background compiling can take up a lot of system resources, so we want
to provide users a way to pause the operation and get some of those CPU
cycles back.

This provides an IdePausable and registers it with the context. The omni
bar can then display the pausable operation in the build popover for
the user to control.

Upon pausing, the current task is cancelled (but work item left on the
queue) and is reprocessed when the operation has been unpaused.