flatpak: work around devhelp
We need access to the documentation that is provided in the flatpak
runtimes as they might not be available on the host.

This cheats and symlinks the gtk-doc/html files from ".Docs" runtimes into
the the data dir used by Builder.

It is very much not an ideal solution, but we need some what to show
documentation to users in the 3.26 cycle and this is about the best we
can do as we're already in code freeze.