notifications: use GNotification
This simplifies the notification plugin to use GNotification and to limit
the number of notifications per-project to one. It also only notifies about
completion if we got to IDE_BUILD_PHASE_BUILD.

It can still update too much for lots of concurrent (but coalesced
requests), but now those all get combined into one, so its effectively
squashed.