fpaste: remove fpaste plugin
This plugin has been broken in a couple different ways, and the
paste server API has completely changed. Instead of trying to
fix this (which requires a bunch), I'd rather just remove this.

If someone wants to make a new version that is well supported,
they can build it however the see fit.