build: remove gnome-builder-cli and gnome-builder-worker
Having these programs split out was not very useful as we move to a
flatpak'd world. If we keep everything accessable via --type= then
we can allow things to work even through 'flatpak run ...'

Since we previously used a "hidden" parameter --type for the plugin
name, that has been renamed to --plugin= which is arguably a much
better choice.

Use --type=cli for the cli version and --type=worker for the plugin
worker version of gnome-builder.

For our old format, you can do something like this in your .bashrc:

Either:

  alias ide="gnome-builder --type=cli"
  alias ide="flatpak run org.gnome.Builder --type=cli"

And then:

  ide --list-commands