gdb: use fullname when possible
We want to try to use fullname when possible, but we need to possibly
translate it from the runtime into a path that we can access outside
of the runtime.

This tries to handle both the relative-to-builddir cases as well as the
case of files found in the SDK headers.