source tree reorganization
As the project has grown, we've let things get a bit out of their
logical place. This does a bit of cleanup and tries to bring some
of the embedded resources closer to where they are used.

But more importantly, this allows us to clean some things up to
the point that we can move forward statically linking a bunch of
the plugins into libide. The plan here is to speed up the initial
loading by avoiding lots of disk I/O stats which are currently
hurting us.