style-guide: use #pragma once header guards
This changes most of the code base to using #pragma once instead
of the old style #ifndef FOO_H header guards. I hear that the
old style is still slightly faster, but I don't care much about
that. I'm sure GCC will continue to optimize for this going
forward.

Also, we're already using a bunch of other C11 features, so
might as well use this too.