gdb: use static library for gdbwire with alternate cflags
If we build the gdbwire amalgamation as a separate static library then we
can alter the CFLAGS that are used to compile it. Given that it is an
amalgamation and is C89-based rather than C11, this allows us to squash
some warnings that are unhelpful.

We link the library with libgnome-builder-plugins.so so the end effect
should be identical to what we had before.