project: use "gio trash" under flatpak
Until we have a working implementation of trash detecting cross-filesystem
boundary moves, we should have a fallback that still tries to trash the
file correctly (without being destructive).

This falls back to "gio trash $uri" on the host. While not ideal, it should
be enough to do the right thing as a stop gap.

https://bugzilla.gnome.org/show_bug.cgi?id=780472