pipeline: protect against spurious builds
If this is an automated build operation, such as those by
fetching build flags, try to protect ourselves against spurious
builds.

A full rebuild or normal build request should still progress as
normal. This simply checks for advancing up to CONFIGURE when
there was a previous failure.

Any UI element (such as clicking Build or activating the
"build" action) will still result in a proper build.