layout: three-finger-swipe to move content
IdeLayoutStack was the basis for our "animated content panes"
which can move an IdeLayoutView between stacks. With a little
bit of work, we can allow performing this operation with a
three-finger-swipe left or right.

This does not use three-finger-swipe for up/down splits as that
is more work than I think is worth the benefit. Although, maybe
someone reading this will work on a patch if they really want
it too.

This makes the content "sticky to fingers" so that it is clear
you are moving the document to the left or right.

This exposes an implementation detail of DzlBoxTheatric in that
it will not clip to the parent window. We might want to add an
option to do that so things look as expected.