vcs: add ide_vcs_path_is_ignored()
This function allows for checking if a file should be ignored
similar to ide_vcs_is_ignored(). It's purpose is to save the
caller from dealing with the burden of checking if a GFile
creation needs to be relative to the project or absolute.

It also is slightly faster in checking against known regexes
for ignored globs.

https://bugzilla.gnome.org/show_bug.cgi?id=774977