testing: implement rudimentary unit testing
Provides basic implementation for unit test plumbing. The UI is
both minimal and incomplete, but enough to land on master so
that future pieces can be implemented later in the cycle.

The test provider is responsible for loading and running the
test. I anticipate that we'll need to tweak the API so that we
can run the tests under gdb. Additionally, we need to track
output so it can be displayed.