buildsystem: translate include paths in build flags
Previously, only the autotools plugin knew how to translate include paths so
that they could be accessed outside the runtime. This caused problems with
build systems like meson when using flatpak because clang did not know how to
access /app (and even some /usr) on the host.

This moves some post-processing to the base class which should allow some
build system plugins to continue being oblivious to the situation.