pipeline: short-circuit when pipeline is past requested phase
If the build pipeline has already past the requested phase, we
can short circuit. This should help in situations where we end
up queuing the request because a build is in progress (which is
beyond the phase we need to complete).

There is still an issue with this where queued items are stuck
behind an higher-requested phase and they will not be flushed
until that build has completed. To fix this issue, we need to
scan forward and flush items with lower requested phases.