pipeline: complete queued tasks before advancing pipeline
If we are advancing the pipeline past the requested phase of any
queued tasks, we want to complete those tasks before processing
the next item.

This improves the situation where the things like requesting
build flags are queued behind a real build, when the necessary
phase has already completed but was queued before the build
started or before it had advanced past the requested phase.