compile-commands: add IdeCompileCommands
This is a helper object to simplify the process of working with
clang-style compile_commands.json databases. Some build systems
such as Meson and CMake can benefit from having access to this
information in a unified manner.

Now that IdeBuildSystem can automatically translate paths from
the runtime build dir, this should allow us to remove some code
from those plugins (and share it in libide instead).

We might still want something to resolve things like -I includes
based on the relative working directory, but in practice, that
should match the $builddir of the build pipeline.