meson: port meson plugin to C
This code is increasingly critical to Builder so I want to ensure that it
is both in a statically typed language and properly implemenst the various
async APIs we expect for interactivity.

This also switches to using IdeCompileCommands to access the build args
without having to advance the build pipeline. Doing so allows us to
potentially avoid running the build machinery immediately upon reopening
a given project. (Something that the Build Pipeline does not handle well
on it's own, and users have been grumpy about).