ctags: try to avoid activating various keywords
We don't have a really great way to deal with this using the
completion system provided by GtkSourceView (as there is not a
way to get the activation key to perform the desired indent or
space). But we can at least avoid activating the item and leave
the insertion cursor where it was.

To do this better, we need to build a new completion engine for
Builder/GSV that does things a bit differently.

Either way, this avoids some common keywords. If you're trying
to get a newline afterwards, you'll have to hit enter a second
time (as this will just dismiss the completion window).

https://bugzilla.gnome.org/show_bug.cgi?id=790101