greeter: add IdeGreeterSection pluggable API
This adds a new IdeGreeterSection interface to be used for
defining sections in the primary greeter. As part of this, the
newcomers section has been removed from libide and will come
back as a plugin.

Additionally, we should be able to remove the recent projects
section from libide and move that to a plugin. That is not part
of this refactoring and can happen at some point in the future,
as it is fairly low priority.