build: vendor GdTaggedEntry as IdeTaggedEntry
We were having a number of issues compiling on systems where libgd is
installed (and not the, typically static, libgd from GNOME). We only use
one widget from there anyway, which doesn't really change. So instead, we
can just vendor it and namespace it into libide. This also reduces the
number of private shared libraries we load by one.

Namespace was changed, and availability macros added using the tool at
https://github.com/chergert/add-availability-macros.

https://bugzilla.gnome.org/show_bug.cgi?id=787578