app: reap legacy data at startup
When starting Builder, start reaping old data created by
legacy versions of Builder. Now that project data is in a
per-project directory, we can just wipe the project directory
when it is deleted.

However, we don't want to leave all this old data around, so
we need to reap it, might as well be now.