sourceview: add IdeSourceScrollAlign
This flags-like enum is created to allow choosing to scroll
X, Y, or both X and Y.

It does not follow the typical mask-style of flags because we
want it to stay compatible with TRUE/FALSE as the parameter
original provided. So we just use a simple macro to determine
if X or Y should be scrolled.