IdeBuildTarget: add get_cwd()
Some build systems (notably, npm), insist that their build targets
be run in a specific working directory, or they'll misbehave.

To allow BuildTargets and BuildTargetProviders specify that,
introduce a vfunc get_cwd(). If this vfunc returns any non-NULL
value, IdeRuntime uses it to set the cwd of the created IdeRunner.

https://bugzilla.gnome.org/show_bug.cgi?id=790846