IdeBuildTarget: add get_language()
Add a way to find the programming language of a build target.
This will be used to weed out incompatible debuggers, profiling
tools, etc.

Cargo build targets are changed to return rust, NPM to return JS.
Similarly, a future Cabal integration should say Haskell, Maven or ANT
should say Java, etc.

For now, automake, meson and cmake return "asm" (native code),
because it's annoying to find if the build target is built from C,
C++, Vala or Fortran, and it does not matter too much because gdb
and valgrind can deal with it.

https://bugzilla.gnome.org/show_bug.cgi?id=790846