IdeDebugManager: check the programming language to see if the debugger is compatible
Make it possible to retrieve the IdeBuildTarget from the IdeRunner,
and use it to retrieve the programming language of the program
being run. Compare it to the list of languages claimed by the
debugger to see if the debugger is compatible.

Tested with npm on host (correctly claims no debugger is available)
and meson on flatpak (correctly starts gdb)

https://bugzilla.gnome.org/show_bug.cgi?id=790846