greeter: use Directory filter by default
Most people seem to be thinking about their projects in terms
of a "project directory". Since we can auto-discover the build
system within that directory, we no longer need to default to
selecting project files. It was at odds with inode/directory
content-type in file-chooser because that requires SELECT_FOLDER
file chooser action.