project-tree: refresh tree when node is expanded
This actually works in the opposite fashion of the byline. What this does
is remove all the children when the node is collapsed, and marks it for
rebuild. That way, it gets rebuilt in the normal fashion when the node is
expanded (and ensures we have a common flow for both cases now).