editor: fix reselection of search text
We had some code to block reselection because the notify callback in the
revealer would reselect the text (causing it delete the previous text).

This instead pushes that check off to the notify callback rather than
in our core grab_focus case.

This fixes reselection with ctrl+f while in the search box.