project-tree: update project tree when file monitor changes
This tries to react to changes in the project tree that are
discovered via inotify (or other backend on non-Linux).

Since the recursive file monitor is particularly suited to
Linux (inotify) in modern GLib versions, we might need to
ratchet things back when we other platforms such as BSD. I'm not
sure how scalable the GFileMonitor backends are there.

To keep things more lightweight, this will not respond to
ignored files by the VCS. I'm not sure its worth the effort to
try to deal with the "show ignored files" option in this case.
It will require waking up lots of monitors that I don't want
to churn handlers upon.

Getting this right could take a lot of tweaks, so if you find
issues (especially with larger projects) do let me know. I
can't imagine adding something of this magnitude will handle
everything on a first try...