util: add non-canonical relative path helper
There may be some cases where we need to get a non-canonical path relative
to a given directory. This tries to handle that by resolving relative paths
from the base file and returning a path with a number of ../ in them.

The driving force for this is that gdb seems to want paths like this in
some cases.