application-actions: Use en dash to split dash range
In the spirit of contributing greatly to the advancement of the GNOME
desktop and its developer tools, I present this remedy for a
typographical flaw found in Builder's About dialog. The range of years
in the copyright date is split using an em dash. That is too much dash.
It should use an en dash instead.

You can't tell the difference from the source code or from diffs,
because you cannot tell the difference with a monospace font. But the en
dash is shorter. Less dashy. Many long arguments could be had on which
type of dash should be used in general, but when it comes to ranges,
there is no controversy. An en it is, an en it must be! For typography!
For Unicode!