project-tree: add DnD plumbing
This doesn't yet use the DzlFileTransfer API to copy/move
things, but it does land the plumbing we need to handle the
node movements.

The other annoying thing is that we are using single-click so
that the treeview feels more like listbox. We need to do some
special case handling so that we don't open the file while we
are really trying to start a drag operation.