project-tree: implement copy and move
This allows dragging contents within the tree as well as from
external programs that support text/uri-list (such as Nautilus).

We rely on the change monitors to rebuild nodes instead of
manually removing nodes from the tree when doing moves. It has
the chance of breaking if the user has run out of inotify
watches, but that's not something I really want to deal with
protecting against here (instead we should get the sysctl bumped
or something to that nature).