project-tree: add tree builder for vcs status
This doesn't yet set the emblems (we need artwork made), but it
will query the VCS status from the given back-end asynchronously
and then update the node with the result.

One thing this doesn't yet handle (that I'd like to) is adding
an emblem to each of the parent nodes. The difficulty currently
is that we would need to synthesize the parent nodes in the
hashtable so that we can add status objects for directories
when querying the VCS.