project-tree: colorize nodes based on vcs status
This is far from perfect, or how we expect things to ship in its
final form. But this does land some plumbing to react to changes
in the VCS layer and ensure that we have a place to update the
tree node information.

I'd like to see something more advanced here, but this gets some
basics in place.