project-tree: simplify vcs tree builder
Now that we can hook into the cell data func from the given
tree builder, we have the ability to avoid setting data directly
on the tree nodes.

Combining this with our vcs monitor, we can ensure we have more
up to date vcs information with much less "tree node tracking"
going on to synthesize folder values (as vcs monitor does that
state tracking for us).