build-manager: improve chained cancellation
This ensures we don't overwrite self->cancellable except when we
cancel the GCancellable ourselves. This is important because we
need to know that our cancellable cannot ever be cancelled
outside of this module (or we'd have to clear it upon
::cancelled).

This fixes an issue where the run manager fails to run a second
time due to a lingering cancellable who has been cancelled.