style: give theme authors a bit of help
There is absolutely *NO GUARANTEE* I will not break your theme on
a daily basis. We allow upstream theming by theme authors that are
willing to be dilligent and keeping things up to date. Builder is
largely built with custom widgets and that requires attentiveness
on the part of those touching the themes.