flatpak: work around flatpak-builder --state-dir bug
It appears that if we don't want updates, flatpak-builder will still query
the remote peer because it fails to find a local mirror directory for the
peer. However, if we specify --extra-sources, we can make it find it as
expected.

We probably need to fix builder_context_find_in_sources_dirs_va() to check
the specified --state-dir.