flatpak: bundle flatpak-builder in our flatpak
This allows us to stay more up-to-date with flatpak-builder, since it knows
how to execute commands on the host using HostCommand.

It also means we don't have to check for flatpak-builder on the host when
bundled in flatpak. So that is one less thing to go wrong on install.

There are a few tricks going on with the environment variables to make sure
that the flatpak commands running on the host find the right installation.